+++ src/Molecules/TabsNav/TabsNav.tsx
@@ -1,11 +1,26 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import styled from 'styled-components';
+import { Component, ItemProps, ScrollStyleProps, TitleProps } from './TabsNav.types';
+import { scrollStyles } from './TabsNav.styles';
+import { useIntersect } from '../../common/Hooks';
 import { Flexbox, TabTitle, Typography } from '../..';
 import { assert } from '../../common/utils';
 import { useKeyboardNavigation } from '../Tabs/useKeyboardNavigation';
-import { Component, ItemProps, TitleComponent } from './TabsNav.types';
 import { useLink } from '../../common/Links';
 import { LinkProps } from '../../common/Links/types';
+
+const DEFAULT_SCROLL_OPTIONS = {
+  active: false,
+  scrollBarHidden: false,
+  scrollIntoViewOptions: null,
+  scrollFade: false,
+};
+
+const DEFAULT_SCROLL_INTO_VIEW_OPTIONS = {
+  behavior: 'smooth',
+  inline: 'center',
+  block: 'end',
+};
 
 export const Item: React.FC<ItemProps> = ({ children }) => {
   return <div>{children}</div>;
@@ -23,41 +38,36 @@ const StyledLink = styled(Link)`
   color: inherit;
 `;
 
-const Title: TitleComponent = ({
-  active,
-  children,
-  setRef,
-  to,
-  onKeyDown,
-  onClick,
-  height,
-  fullServerRedirect,
-}) => {
-  return (
-    <Typography type="primary" weight={active ? 'bold' : 'regular'}>
-      <StyledLink
-        to={to}
-        innerRef={setRef}
-        aria-current={active ? 'page' : undefined}
-        onKeyDown={onKeyDown}
-        onClick={onClick}
-        fullServerRedirect={fullServerRedirect}
-      >
-        <TabTitle active={active} height={height}>
-          {children}
-        </TabTitle>
-      </StyledLink>
-    </Typography>
-  );
-};
-Title.displayName = 'TabsNav.Title';
-
-const StyledUl = styled.ul`
+const StyledFlexbox = styled(Flexbox)<ScrollStyleProps>`
   list-style-type: none;
   padding: 0;
   margin-top: 0;
   margin-bottom: 0;
+  ${scrollStyles}
 `;
+
+const Title = React.forwardRef<HTMLSpanElement, TitleProps>(
+  ({ active, children, setRef, to, onKeyDown, onClick, height, fullServerRedirect }, ref) => {
+    return (
+      // @ts-expect-error: Typography doesn't accept ref but is actually a React.ForwardRef
+      <Typography type="primary" weight={active ? 'bold' : 'regular'} ref={ref}>
+        <StyledLink
+          to={to}
+          innerRef={setRef}
+          aria-current={active ? 'page' : undefined}
+          onKeyDown={onKeyDown}
+          onClick={onClick}
+          fullServerRedirect={fullServerRedirect}
+        >
+          <TabTitle active={active} height={height} variant="large">
+            {children}
+          </TabTitle>
+        </StyledLink>
+      </Typography>
+    );
+  },
+);
+Title.displayName = 'TabsNav.Title';
 
 const isItemOrUndefined = (x: any): x is { type: typeof Item; props: ItemProps } => {
   if (x == null || typeof x === 'undefined') {
@@ -67,18 +77,52 @@ const isItemOrUndefined = (x: any): x is { type: typeof Item; props: ItemProps }
   return typeof x === 'object' && Object.hasOwnProperty.call(x, 'type'); // FIXME: && x.type === Item;
 };
 
-export const TabsNav: Component = ({ children, height = 11, className }) => {
+export const TabsNav: Component = ({
+  children,
+  height = 11,
+  className,
+  scrollOptions = DEFAULT_SCROLL_OPTIONS,
+}) => {
   const { setRef, onKeyDown } = useKeyboardNavigation({
     itemsLength: React.Children.count(children),
   });
   const titles: React.ReactNode[] = [];
+
+  const [setIntersectionLeftRef, intersectionLeftRatio] = useIntersect<HTMLDivElement>();
+  const [setIntersectionRightRef, intersectionRightRatio] = useIntersect<HTMLDivElement>();
+
+  const scrollToRefCallback = useCallback(
+    (ref) => {
+      if (ref && scrollOptions.active) {
+        setTimeout(() => {
+          ref.scrollIntoView(
+            scrollOptions.scrollIntoViewOptions ?? DEFAULT_SCROLL_INTO_VIEW_OPTIONS,
+          );
+        }, 0);
+      }
+    },
+    [scrollOptions.active, scrollOptions.scrollIntoViewOptions],
+  );
+
+  const setIntersectionRef = (index: number) => {
+    if (!scrollOptions.active) {
+      return null;
+    }
+    if (index === 0) {
+      return setIntersectionLeftRef;
+    }
+    if (index === React.Children.count(children) - 1) {
+      return setIntersectionRightRef;
+    }
+    return null;
+  };
 
   React.Children.forEach(children, (c, i) => {
     if (!isItemOrUndefined(c)) {
       assert(false, 'There should be only <TabsNav.Tab> children inside of <TabsNav> component');
     } else if (c) {
       titles.push(
-        <Flexbox item as="li" key={c.props.to}>
+        <Flexbox ref={setIntersectionRef(i)} item as="li" key={c.props.to}>
           <Title
             active={!!c.props.active}
             onClick={c.props.onTitleClick}
@@ -88,6 +132,7 @@ export const TabsNav: Component = ({ children, height = 11, className }) => {
             onKeyDown={onKeyDown}
             height={height}
             className={c.props.className}
+            ref={c.props.active ? scrollToRefCallback : null}
           >
             {c.props.title}
           </Title>
@@ -97,16 +142,18 @@ export const TabsNav: Component = ({ children, height = 11, className }) => {
   });
 
   return (
-    <Flexbox
+    <StyledFlexbox
+      $height={height}
+      $scrollOptions={scrollOptions}
+      $intersectionLeftRatio={intersectionLeftRatio || 0}
+      $intersectionRightRatio={intersectionRightRatio || 0}
       className={className}
       container
-      direction="row"
-      gutter={4}
-      as={StyledUl}
-      sm={{ gutter: 8 }}
     >
-      {titles}
-    </Flexbox>
+      <Flexbox container direction="row" gutter={4} sm={{ gutter: 8 }}>
+        {titles}
+      </Flexbox>
+    </StyledFlexbox>
   );
 };