+++ src/Molecules/TabsNav/TabsNav.types.ts
@@ -1,3 +1,5 @@
+export type HtmlProps = Omit<React.HTMLProps<HTMLSpanElement>, 'children'>;
+
 export type ItemProps = {
   to: string;
   fullServerRedirect?: boolean;
@@ -7,17 +9,22 @@ export type ItemProps = {
   className?: string;
   active?: boolean;
 };
+export type ItemComponent = React.FC<ItemProps>;
 
 export type ContainerProps = {
   height?: number;
   className?: string;
+  children?: React.ReactNode;
+  scrollOptions?: {
+    active: boolean;
+    scrollBarHidden: boolean;
+    scrollIntoViewOptions?: { behavior?: string; inline?: string; block?: string };
+    scrollFade?: boolean;
+  };
 };
-
-export type ItemComponent = React.FC<ItemProps>;
 export type Component = React.FC<ContainerProps> & { Tab: ItemComponent };
 
-export type TitleComponent = React.FC<TitleProps>;
-export type TitleProps = {
+export type TitleProps = HtmlProps & {
   active: boolean;
   className?: string;
   onClick?: React.MouseEventHandler;
@@ -27,4 +34,11 @@ export type TitleProps = {
   onKeyDown: (e: React.KeyboardEvent) => void;
   height?: number;
   fullServerRedirect?: boolean;
+};
+
+export type ScrollStyleProps = {
+  $height: number;
+  $scrollOptions: { active: boolean; scrollBarHidden: boolean; scrollFade?: boolean };
+  $intersectionLeftRatio?: number;
+  $intersectionRightRatio?: number;
 };